Fixed Gigantic Memory Leak
So, turns out when you delete an object with a mesh, unity still keeps a
reference to the mesh or something.  None of our lists were getting too
big, Unity was keeping too many references.  Deleting
MeshFilter.sharedMesh removes the reference though, so I adding that in
fixed the problem.