Made more improvements to the A* algorithm
Before, when the AI picked its starting node, it would just pick the
closest one. Now it picks the node that would creates the most efficient
path.
Added destruction of start node and destination node after the algorithm
finished.
A* function now returns a list of vector2 that represent the path rather
than the nodes itself, because when you delete the start node and the
destination node, the path list would be missing those 2 nodes.