FIxed node cleanup on A* algorithm.
Previously the A* algorithm instantiated two extra game objects that
wasn't being used and wasn't cleaned up properly.
Also the enemy now pathfinds in real time in the update loop, instead of
once during the start. Runs surprisingly well, and I tested this with 50
baddies chasing after the goal.
To try this out, just drag the GOAL around and watch the enemy chases
after it. Note that there is a bug where if you drag the goal to
somewhere the enemy can't reach then it'll error.